Move shared types to Miso.Html.Types module
